ZON-5567: Fold query descriptor into configuration module where it is used
